FIX: Refactor and fix multiapi
