+++ indicoio/image/image_recognition.py
@@ -1,6 +1,6 @@
-from indicoio.utils.image import image_preprocess
-from indicoio.utils.api import api_handler
-from indicoio.utils.decorators import detect_batch_decorator
+from ..utils.image import image_preprocess
+from ..utils.api import api_handler
+from ..utils.decorators import detect_batch_decorator
 
 
 @detect_batch_decorator
