+++ indicoio/multi/__init__.py
@@ -1,0 +1,3 @@
+from .analyze_text import analyze_text
+from .analyze_image import analyze_image
+from .intersections import intersections
