+++ indicoio/text/sentiment_hq.py
@@ -2,7 +2,7 @@ from ..utils.api import api_handler
 from ..utils.decorators import detect_batch_decorator
 
 @detect_batch_decorator
-def sentiment(text, cloud=None, batch=False, api_key=None, version=None, **kwargs):
+def sentiment_hq(text, cloud=None, batch=False, api_key=None, version=None, **kwargs):
     """
     Given input text, returns a scalar estimate of the sentiment of that text.
     Values are roughly in the range 0 to 1 with 0.5 indicating neutral sentiment.
@@ -12,15 +12,15 @@ def sentiment(text, cloud=None, batch=False, api_key=None, version=None, **kwarg
 
     .. code-block:: python
 
-       >>> from indicoio import sentiment
+       >>> from indicoio import sentimenthq
        >>> text = 'Thanks everyone for the birthday wishes!! It was a crazy few days ><'
-       >>> sentiment = sentiment(text)
+       >>> sentiment = sentimenthq(text)
        >>> sentiment
-       0.6946439339979863
+       0.6210052967071533
 
     :param text: The text to be analyzed.
     :type text: str or unicode
     :rtype: Float
     """
     url_params = {"batch": batch, "api_key": api_key, "version": version}
-    return api_handler(text, cloud=cloud, api="sentiment", url_params=url_params, **kwargs)
+    return api_handler(text, cloud=cloud, api="sentimenthq", url_params=url_params, **kwargs)
