+++ tests/test_batching.py
@@ -11,7 +11,7 @@ class TestBatchSize(unittest.TestCase):
         self.api_key = config.api_key
 
         if not self.api_key:
-            raise SkipTest
+            raise unittest.SkipTest
 
     def test_batch_size(self):
         test_data = ["Terribly interesting test data."] * 100
@@ -26,13 +26,14 @@ class TestBatchSize(unittest.TestCase):
             response = sentiment_hq(test_data, batch_size=20)
 
         files = glob.glob('indico-sentimenthq-*.json')
+        print files
         assert len(files)
 
-        for file in files:
-            data = json.load(open(file))
+        for filename in files:
+            data = json.load(open(filename, 'wb'))
 
             # first four batches should have returned
             assert len(data) == 80
 
             # clean up after ourselves
-            os.remove(file)
+            os.remove(filename)
